Add option to handle absent cross_az_attach
At the moment volumes for workers are gonna be spawn in a default
AZ completely disregarding user request for Availability Zone.

In a design where cross_az_attach is disabled, an attempt
to add a volume to worker has high failure percentage due to fallback
to the default scheduling zone, unless allow_availability_zone_fallback
is disabled.

This patch adds a configuration option `cross_az_attach` which is set
to True by default to align with Nova typical behavior. It will define
AZ to be set to `nova` according to a CSI default [1].

When `cross_az_attach` is set to False, AZ for the volume will be set to
the cluster AZ value.
This ensures that volume will be created in a same zone as workers are
preventing failures.

[1] https://github.com/kubernetes/cloud-provider-openstack/blob/d228854cf58e7b4ed93d5e7ba68ab639450e3449/docs/cinder-csi-plugin/using-cinder-csi-plugin.md#supported-parameters

Relates-to: #366